datahike-postgres uses new configuration format and supports ssl
I added the support for ssl-connections to this library and due
to changes to datahike config and simplicity I chose to switch
the configuration format to map from now on. This change to
datahike-postgres will be breaking so it is needed to be released
in combination and updated combined with datahike itself.

- Closes https://github.com/replikativ/datahike/issues/133
- Refers #2
- Breaking Change